requirements.txt
python-dotenv==1.0.0
mysql-connector-python==8.2.0
requests==2.31.0
pytest==7.4.3
schedule==1.2.1
pytest-mock==3.12.0 